Rename withAuthUserTokenSSR file
